initial assessment fix
